backend/requirements.txt
Django
djangorestframework
psycopg[binary]
gunicorn
python-dotenv
django-cors-headers
uvicorn
django-environ